+++ src/Geta.EPi.Find.Extensions/TypeSearchExtensions.cs
@@ -18,6 +18,7 @@ namespace Geta.EPi.Find.Extensions
     public static class TypeSearchExtensions
     {
         private static readonly ILogger Log = LogManager.GetLogger(typeof(TypeSearchExtensions));
+
         /// <summary>
         /// Add a filter conditionally, makes it easier to write a fluent query.
         /// </summary>
@@ -26,6 +27,7 @@ namespace Geta.EPi.Find.Extensions
         /// <param name="condition">if set to <c>true</c> the filterExpression will be added.</param>
         /// <param name="filterExpression">The filter expression.</param>
         /// <returns>Updated search.</returns>
+        [Obsolete("Use the Conditional method instead")]
         public static ITypeSearch<TSource> ConditionalFilter<TSource>(
             this ITypeSearch<TSource> search, bool condition, Expression<Func<TSource, Filter>> filterExpression)
         {
@@ -35,6 +37,24 @@ namespace Geta.EPi.Find.Extensions
             }
 
             return search.Filter<TSource>(filterExpression);
+        }
+
+        /// <summary>
+        /// Add a filter expression if the condition is true
+        /// </summary>
+        /// <typeparam name="TSource">The type of the source.</typeparam>
+        /// <param name="search">The search.</param>
+        /// <param name="condition">if set to <c>true</c> the filterExpression will be added.</param>
+        /// <param name="request">The filter expression.</param>
+        /// <returns>Updated search.</returns>
+        public static ITypeSearch<TSource> Conditional<TSource>(this ITypeSearch<TSource> search, bool condition,
+            Func<ITypeSearch<TSource>, ITypeSearch<TSource>> request)
+        {
+            if (!condition)
+            {
+                return search;
+            }
+            return request(search);
         }
 
         /// <summary>